[FIX] CheckpointSpawn / Audio / Colliders / pause
[Fix]

Checkpoints previously kept the player at the checked point even when the game was fully reset, I have now modified it in a way that at the start of the game it modified the value from true to false, so that it is affected to only go back to the current checkpoint, and if the game is restarted again at the start the value is true and it adds the initial position of the player.

Audio Fixed it and readded it to the scene, it was on some object in game that was not visable previously and with that got removed, which is now added back as standalone.

Colliders now the game is not epossible to end and you cant jump on blocks, previously some areas the player would be stuck forever with no way of continuing from it which is now simplified addressed.

Fixed Pause Menu to instead go to the main Menu, rather than quit the game.
Changed time.scale to 1 and 0 depending if the pause menu is active

[ADDS]
Adds playerPrefs feature so they can be seen if required..

Added a gameIcon for the build splash screen.

Changed the project currently for Web support.