Unlockables temp to true/ code related issues address
Some of the scripts and changes are still Only within the Stuart folder, but will be kept there as currently required functionalities of the game have been addressed.

Changed Dash and Shooting to be temporary True
as unlockables are not added in the game so by default it will be true.

Fixed some crashes